md5sum.c: add a new implementation from Gnulib
This commit adds a fresh new MD5 implementation from
Gnulib. This implementation is well-tested, clean, easy to use,
and license-compatible.

This implementation brings clean code
with nearly no performance overhead, just less a second slower
than the previous implementation from BSD.

Gnulib is the portable standard library from GNU,
and is used in many GNU projects. Actually, "md5sum" command
from coreutils is using this implementation.

Since it's a copy of algorithm from Gnulib, we do not need to maintain it.
It is not recommend to "optimize" or "refactor" it for no reason.

Signed-off-by: Tom Li <biergaizi@member.fsf.org>